Changes the location of the function
